feat: Major game systems implementation and content expansion
- Add CareerManager for job search, salary, promotion system
- Add RelationshipManager for social interactions, dating, marriage
- Add EducationManager for school, YKS exam, university system
- Fix critical serialization issues (decimal -> long for JSON)
- Fix HashSet FIFO issue with Queue for recent events
- Integrate all new systems with GameManager yearly updates
- Add 15+ new diverse events (crime, health, travel, business)
- Add Asset/Debt system for financial tracking
- Update GameBootstrap with new manager initialization

+++ Assets/Scripts/Core/EventBus.cs
@@ -164,9 +164,9 @@ namespace TurkishLifeSim.Core
     /// </summary>
     public struct MoneyChangedEvent : IGameEvent
     {
-        public decimal OldAmount;
-        public decimal NewAmount;
-        public decimal Delta;
+        public long OldAmount;
+        public long NewAmount;
+        public long Delta;
         public string Reason;
     }
